Makefile: Add initial bootstrapping!
I also have the beginnings of building the ports tree, but that has
proven to be more involved than I wanted it to be.  For now we're doing
an "untrusted" pull of the Alpine keys off of the Alpine HTTPS
repository then doing key signing for everything else.  I plan on doing
that phase from the ports tree at a later date, but for now it has
proven too elusive...